+++ skills/declarative/extensions/Microsoft.Bot.Solutions.Extensions/HttpClient.cs
@@ -4,13 +4,13 @@ using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
 
-namespace Microsoft.Bot.Solutions.Extensions.Services
+namespace Microsoft.Bot.Solutions.Extensions
 {
-    public static class HttpClientManager
+    public static class HttpClient
     {
-        public static HttpClient GetAuthenticatedClient(string accessToken)
+        public static System.Net.Http.HttpClient GetAuthenticatedClient(string accessToken)
         {
-            var httpClient = new HttpClient();
+            var httpClient = new System.Net.Http.HttpClient();
             httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", accessToken);
             return httpClient;
         }
