+++ skills/declarative/extensions/Microsoft.Bot.Solutions.Extensions/Schemas/Actions/Microsoft.Graph.Who.GetDirectReports.schema
