+++ skills/declarative/extensions/Microsoft.Bot.Solutions.Extensions/Schemas/Actions/Microsoft.Graph.Who.GetUsers.schema
