+++ skills/declarative/whoskill/dialogs/FilterResult/FilterResult.dialog
@@ -1,0 +1,46 @@
+{
+  "$kind": "Microsoft.AdaptiveDialog",
+  "$designer": {
+    "id": "QfqKSV",
+    "name": "FilterResult"
+  },
+  "autoEndDialog": true,
+  "defaultResultProperty": "dialog.result",
+  "triggers": [
+    {
+      "$kind": "Microsoft.OnBeginDialog",
+      "$designer": {
+        "name": "BeginDialog",
+        "id": "QXjAca"
+      },
+      "actions": [
+        {
+          "$kind": "Microsoft.Graph.Who.GetUsers",
+          "token": "=user.token.token",
+          "keywordProperty": "=dialog.chosenPerson.Mail",
+          "resultProperty": "dialog.users",
+          "$designer": {
+            "id": "Xo1zQ1"
+          },
+          "topProperty": "1"
+        },
+        {
+          "$kind": "Microsoft.SetProperty",
+          "$designer": {
+            "id": "tatxPv"
+          },
+          "property": "dialog.persons",
+          "value": "=where(dialog.persons, x, x.Id != dialog.users[0].Id)"
+        },
+        {
+          "$kind": "Microsoft.EndDialog",
+          "$designer": {
+            "id": "fkV12N"
+          },
+          "value": "=dialog.persons"
+        }
+      ]
+    }
+  ],
+  "generator": "FilterResult.lg"
+}
