+++ skills/declarative/whoskill/dialogs/FilterResult/language-generation/en-us/FilterResult.en-us.lg
@@ -1,0 +1,1 @@
+[import](common.lg)
