+++ skills/declarative/whoskill/dialogs/FormatPersonModel/language-generation/en-us/FormatPersonModel.en-us.lg
@@ -1,0 +1,4 @@
+[import](common.lg)
+
+
+
