+++ skills/declarative/whoskill/dialogs/ProcessSearchResult/language-generation/en-us/ProcessSearchResult.en-us.lg
@@ -8,3 +8,4 @@
 - ${NoSearchResult(conversation.luisResult.keyword)}
 
 
+
