--- skills/declarative/whoskill/extensions/Microsoft.Bot.Solutions.Extensions/Actions/GetMe.cs
@@ -1,55 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
-using AdaptiveExpressions.Properties;
-using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.TraceExtensions;
-using Microsoft.Bot.Solutions.Extensions.Models;
-using Microsoft.Bot.Solutions.Extensions.Services;
-using Newtonsoft.Json;
-
-namespace Microsoft.Bot.Solutions.Extensions.Actions
-{
-    public class GetMe : Dialog
-    {
-        [JsonProperty("$kind")]
-        public const string DeclarativeType = "Microsoft.Graph.Who.GetMe";
-
-        [JsonConstructor]
-        public GetMe([CallerFilePath] string callerPath = "", [CallerLineNumber] int callerLine = 0)
-            : base()
-        {
-            this.RegisterSourceLocation(callerPath, callerLine);
-        }
-
-        [JsonProperty("resultProperty")]
-        public string ResultProperty { get; set; }
-
-        [JsonProperty("token")]
-        public StringExpression Token { get; set; }
-
-        public override async Task<DialogTurnResult> BeginDialogAsync(DialogContext dc, object options = null, CancellationToken cancellationToken = default)
-        {
-            var dcState = dc.State;
-            var token = this.Token.GetValue(dcState);
-
-            var result = await GraphService.GetCurrentUser(token);
-
-            var currentUser = new WhoSkillUser(token, result);
-
-            // Write Trace Activity for the http request and response values
-            await dc.Context.TraceActivityAsync(nameof(GetMe), currentUser, valueType: DeclarativeType, label: this.Id).ConfigureAwait(false);
-
-            if (this.ResultProperty != null)
-            {
-                dcState.SetValue(ResultProperty, currentUser);
-            }
-
-            // return the actionResult as the result of this operation
-            return await dc.EndDialogAsync(result: currentUser, cancellationToken: cancellationToken);
-        }
-    }
-}
