--- skills/declarative/whoskill/extensions/Microsoft.Bot.Solutions.Extensions/Actions/GetPeers.cs
@@ -1,81 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Runtime.CompilerServices;
-using System.Threading;
-using System.Threading.Tasks;
-using AdaptiveExpressions.Properties;
-using Microsoft.Bot.Builder.Dialogs;
-using Microsoft.Bot.Builder.TraceExtensions;
-using Microsoft.Bot.Solutions.Extensions.Models;
-using Microsoft.Bot.Solutions.Extensions.Services;
-using Microsoft.Graph;
-using Newtonsoft.Json;
-
-namespace Microsoft.Bot.Solutions.Extensions.Actions
-{
-    public class GetPeers : Dialog
-    {
-        [JsonProperty("$kind")]
-        public const string DeclarativeType = "Microsoft.Graph.Who.GetPeers";
-
-        [JsonConstructor]
-        public GetPeers([CallerFilePath] string callerPath = "", [CallerLineNumber] int callerLine = 0)
-            : base()
-        {
-            this.RegisterSourceLocation(callerPath, callerLine);
-        }
-
-        [JsonProperty("resultProperty")]
-        public string ResultProperty { get; set; }
-
-        [JsonProperty("token")]
-        public StringExpression Token { get; set; }
-
-        [JsonProperty("idProperty")]
-        public StringExpression IdProperty { get; set; }
-
-        public override async Task<DialogTurnResult> BeginDialogAsync(DialogContext dc, object options = null, CancellationToken cancellationToken = default)
-        {
-            var dcState = dc.State;
-            var token = this.Token.GetValue(dcState);
-            var idProperty = this.IdProperty.GetValue(dcState);
-
-            var manager = await GraphService.GetManager(token, idProperty);
-            if (manager == null)
-            {
-                if (this.ResultProperty != null)
-                {
-                    dcState.SetValue(ResultProperty, null);
-                }
-
-                return await dc.EndDialogAsync(result: null, cancellationToken: cancellationToken);
-            }
-
-            var directReports = await GraphService.GetDirectReports(token, (manager as User).Id);
-            var currentUser = await GraphService.GetUser(token, idProperty);
-            var peers = new List<WhoSkillUser>();
-            foreach (var user in directReports)
-            {
-                if (currentUser.Id == (user as User).Id)
-                {
-                    continue;
-                }
-
-                peers.Add(new WhoSkillUser(token, user as User));
-            }
-
-            // Write Trace Activity for the http request and response values
-            await dc.Context.TraceActivityAsync(nameof(GetPeers), peers, valueType: DeclarativeType, label: this.Id).ConfigureAwait(false);
-
-            if (this.ResultProperty != null)
-            {
-                dcState.SetValue(ResultProperty, peers);
-            }
-
-            // return the actionResult as the result of this operation
-            return await dc.EndDialogAsync(result: peers, cancellationToken: cancellationToken);
-        }
-    }
-}
