--- skills/declarative/whoskill/extensions/Microsoft.Bot.Solutions.Extensions/Microsoft.Bot.Solutions.Extensions.csproj
@@ -1,12 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Bot.Builder.Dialogs.Adaptive" Version="4.9.0-preview-200504-127776" />
-    <PackageReference Include="Microsoft.Graph" Version="3.0.0" />
-  </ItemGroup>
-
-</Project>
