+++ skills/declarative/whoskill/runtime/ComposerRuntime.sln
@@ -9,7 +9,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.BotFramework.Comp
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.BotFramework.Composer.Functions", "azurefunctions\Microsoft.BotFramework.Composer.Functions.csproj", "{9FC308C1-F0EC-41BE-90FD-9DC00DA7B80B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Bot.Solutions.Extensions", "..\extensions\Microsoft.Bot.Solutions.Extensions\Microsoft.Bot.Solutions.Extensions.csproj", "{03288C2A-7D2A-4A4E-A6BB-98496E7291B4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.Bot.Solutions.Extensions", "..\..\extensions\Microsoft.Bot.Solutions.Extensions\Microsoft.Bot.Solutions.Extensions.csproj", "{A8DB7A9C-B305-4DFD-8E10-F3D8CB2C70B0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -29,10 +29,10 @@ Global
 		{9FC308C1-F0EC-41BE-90FD-9DC00DA7B80B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9FC308C1-F0EC-41BE-90FD-9DC00DA7B80B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{9FC308C1-F0EC-41BE-90FD-9DC00DA7B80B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{03288C2A-7D2A-4A4E-A6BB-98496E7291B4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{03288C2A-7D2A-4A4E-A6BB-98496E7291B4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{03288C2A-7D2A-4A4E-A6BB-98496E7291B4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{03288C2A-7D2A-4A4E-A6BB-98496E7291B4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{A8DB7A9C-B305-4DFD-8E10-F3D8CB2C70B0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A8DB7A9C-B305-4DFD-8E10-F3D8CB2C70B0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A8DB7A9C-B305-4DFD-8E10-F3D8CB2C70B0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A8DB7A9C-B305-4DFD-8E10-F3D8CB2C70B0}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
